refactor(git): extract cross-family helpers into _git_shared.py
