+++ .rrt/tree.lock.toml
@@ -1,10 +1,10 @@
 [meta]
-generated_at = "2026-06-26T12:51:58+00:00"
-rrt_version = "1.10.0"
+generated_at = "2026-07-05T07:31:39+00:00"
+rrt_version = "1.10.1"
 
 [snapshot]
-entry_count = 320
-tree_hash = "sha256:91068c98f4118c291820e3a06440e9fcf8e5b310d4f5a00294b16a2c644fd2e7"
-updated_at = "2026-06-26T12:51:58+00:00"
-ignored_count = 205
+entry_count = 321
+tree_hash = "sha256:87bce5185b25a03862dc918990d2e3a60d95b3d5606005f5eec0ab3b5e7e16f0"
+updated_at = "2026-07-05T07:31:39+00:00"
+ignored_count = 32
 phantom_empty_dirs = []
